refactor(components): reorganize into feature-based folders
Phase 4 of frontend architecture refactor. Moves the remaining 10
flat components from src/components/ into their feature folders:

- common/: RecordingCard, AudioVisualizer, ApiConfigDialog, UpdateModal,
  DashboardHeader (moved in Phase 3)
- dashboard/transcription/: TranscriptionList, TranscriptionDetails
- notes/: NotesTab, AiActionMenu
- logs/: LogsTab
- settings/: UpdaterTab

Six import paths updated across Dashboard.tsx, HistoriqueTab.tsx,
NotesEditorFooter.tsx, TranscriptionSection.tsx, UpdaterSection.tsx,
and RecordingCard.tsx. UpdateModal and UpdaterTab also had relative
./ui/ imports fixed to @/components/ui/ after the move.

No behavior change. Progress tracked in docs/REFACTOR_FRONTEND_PROGRESS.md.
