+++ src/components/common/AudioVisualizer.tsx
