+++ src/components/dashboard/transcription/TranscriptionDetails.tsx
