+++ src/components/logs/LogsTab.tsx
